requirements.txt
opencv-python==4.12.0.88
tflite-runtime==2.14.0
numpy==2.2.6
RPi.GPIO==0.7.1
RPLCD==1.4.0
Adafruit_DHT==1.4.0
